[add] Azure File model [remove] File Upload component [optimize] upgrade Upstream packages
