+++ components/work/WorkEdit.tsx
@@ -1,12 +1,12 @@
 import { observer } from 'mobx-react';
+import { FileUploader } from 'mobx-restful-table';
 import { FormEvent, PureComponent } from 'react';
-import { Container } from 'react-bootstrap';
-import { Button, Col, Form, Row } from 'react-bootstrap';
+import { Button, Col, Container, Form, Row } from 'react-bootstrap';
 import { formToJSON } from 'web-utility';
 
 import activityStore from '../../models/Activity';
+import fileStore from '../../models/File';
 import { i18n } from '../../models/Translation';
-import { FileUpload } from '../FileUpload';
 
 const { t } = i18n;
 
@@ -127,7 +127,8 @@ export class WorkEdit extends PureComponent<WorkEditProps> {
                 {t('upload_file')}
               </Form.Label>
               <Col sm={10}>
-                <FileUpload
+                <FileUploader
+                  store={fileStore}
                   accept="video/*,image/*,.pdf,.doc,.docx,.xml,application/msword,application/vnd.openxmlformats-officedocument.wordprocessingml.document"
                   name="url"
                   max={1}
