+++ package.json
@@ -24,7 +24,7 @@
     "classnames": "^2.3.2",
     "echarts-jsx": "^0.5.0",
     "editorjs-html": "^3.4.2",
-    "idea-react": "^0.27.11",
+    "idea-react": "^1.0.0-beta.1",
     "koajax": "^0.8.4",
     "leaflet": "^1.9.4",
     "leaflet.chinatmsproviders": "^3.0.6",
@@ -33,7 +33,7 @@
     "mobx-i18n": "^0.3.14",
     "mobx-react": "^6.3.1",
     "mobx-restful": "^0.6.4",
-    "mobx-restful-table": "^0.10.0",
+    "mobx-restful-table": "^0.11.0",
     "next": "^12.3.4",
     "react": "^17.0.2",
     "react-bootstrap": "^2.7.4",
@@ -44,20 +44,20 @@
     "web-utility": "^4.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.22.0",
+    "@babel/core": "^7.22.1",
     "@editorjs/editorjs": "^2.27.0",
     "@editorjs/paragraph": "^2.9.0",
     "@octokit/openapi-types": "^17.2.0",
     "@types/leaflet": "^1.9.3",
     "@types/lodash": "^4.14.195",
-    "@types/node": "^16.18.33",
-    "@types/react": "^17.0.59",
+    "@types/node": "^16.18.34",
+    "@types/react": "^17.0.60",
     "eslint": "^8.41.0",
     "eslint-config-next": "^12.3.4",
     "eslint-plugin-simple-import-sort": "^10.0.0",
     "husky": "^8.0.3",
     "less": "^4.1.3",
-    "less-loader": "^11.1.0",
+    "less-loader": "^11.1.1",
     "lint-staged": "^13.2.2",
     "next-pwa": "^5.6.0",
     "next-with-less": "^2.0.5",
